Added utils to seperate out calculations.
